enhance/requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==1.10
django-fcm==0.1.1
django-filter==1.1.0
django-qrcode==0.3
django-widget-tweaks==1.4.1
djangorestframework==3.7.7
idna==2.6
Markdown==2.6.11
mysqlclient==1.3.12
Pillow==5.0.0
pytz==2018.3
reportlab==3.4.0
requests==2.18.4
urllib3==1.22
